add new product and upload image with cloudinary

+++ pages/api/products/[id].js
@@ -6,7 +6,9 @@ export default async function handler(req, res) {
     const {
         method,
         query: { id },
+        cookies
     } = req;
+    const token = cookies.token
 
     dbConnect();
 
@@ -22,6 +24,9 @@ export default async function handler(req, res) {
 
     // update product
     if (method === "PUT") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("Not authenticated!")
+        }
         try {
             const product = await Product.findByIdAndUpdate(id, req.body, {
                 new: true,
@@ -35,6 +40,9 @@ export default async function handler(req, res) {
 
     // delete product
     if (method === "DELETE") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("Not authenticated!")
+        }
         try {
             const product = await Product.findByIdAndDelete(id);
             res.status(200).json("the product deleted successfully");
